fim da correções projeto iniciando

+++ src/main/resources/db/migration/V2__create_city.sql
@@ -1,7 +1,7 @@
 create table city (
   id serial not null,
   name varchar(100) not null,
-  id_state int8 not null,
+  id_state bigint not null,
   constraint city_pk primary key (id),
   constraint state_fk foreign key (id_state) references state(id)
 );
