+++ src/main/resources/db/migration/V3__create_address.sql
@@ -2,7 +2,7 @@ create table address (
   id serial not null,
   street varchar(100) not null,
   cep varchar (10),
-  id_city int8 not null,
+  id_city bigint not null,
   constraint address_pk primary key (id),
   constraint city_fk foreign key (id_city) references city(id)
 );
